refactor(FFI): move all FFI related functionality to its own module
