+++ crates/bls-crypto/src/bls/mod.rs
@@ -1,6 +1,4 @@
 /// Implements BLS signatures as specified in https://crypto.stanford.edu/~dabo/pubs/papers/BLSmultisig.html.
-pub mod ffi;
-
 mod secret;
 pub use secret::PrivateKey;
