+++ crates/bls-crypto/src/lib.rs
@@ -2,8 +2,8 @@
 //!
 //! This crate implements cryptographic operations for BLS signatures
 /// BLS signing
-mod bls;
-pub use bls::{ffi, PrivateKey, PublicKey, PublicKeyCache, Signature};
+pub(crate) mod bls;
+pub use bls::{PrivateKey, PublicKey, PublicKeyCache, Signature};
 
 /// Hashing to curve utilities
 pub mod hash_to_curve;
@@ -13,31 +13,27 @@ pub use hash_to_curve::HashToCurve;
 pub mod hashers;
 pub use hashers::XOF;
 
-use ffi::{Message, MessageFFI};
-use hash_to_curve::try_and_increment::{COMPOSITE_HASH_TO_G1, DIRECT_HASH_TO_G1};
+#[cfg(feature = "ffi")]
+pub mod ffi;
 
-use algebra::{
-    bls12_377::{Fq, Fq2, G1Affine, G2Affine},
-    AffineCurve, FromBytes, ProjectiveCurve, ToBytes,
-};
+#[cfg(any(test, feature = "test-helpers"))]
+pub mod test_helpers;
+
 use log::error;
-use once_cell::sync::Lazy;
-use rand::thread_rng;
-use std::{error::Error, fmt::Display, os::raw::c_int, slice};
 use thiserror::Error;
 
 /// Convenience result alias
 pub type BlsResult<T> = std::result::Result<T, BLSError>;
 
 /// Domain separator for signing messages
-pub static SIG_DOMAIN: &[u8] = b"ULforxof";
-/// Domain separator for Proofs of Posession
-pub static POP_DOMAIN: &[u8] = b"ULforpop";
-/// Domain separator for public inputs to the snark
-pub static OUT_DOMAIN: &[u8] = b"ULforout";
+pub const SIG_DOMAIN: &[u8] = b"ULforxof";
 
-#[cfg(any(test, feature = "test-helpers"))]
-pub mod test_helpers;
+/// Domain separator for Proofs of Posession
+pub const POP_DOMAIN: &[u8] = b"ULforpop";
+
+/// Domain separator for public inputs to the snark
+pub const OUT_DOMAIN: &[u8] = b"ULforout";
+
 #[derive(Debug, Error)]
 /// Error type
 pub enum BLSError {
@@ -57,500 +53,4 @@ pub enum BLSError {
     HashToCurveError,
     #[error("{0}")]
     SerializationError(#[from] algebra::SerializationError),
-}
-
-#[no_mangle]
-/// Initializes the lazily evaluated hashers.
-pub extern "C" fn init() {
-    Lazy::force(&COMPOSITE_HASH_TO_G1);
-    Lazy::force(&DIRECT_HASH_TO_G1);
-}
-
-/// # Safety
-///
-/// out_private_key must initialized to memory that can contain a pointer.
-#[no_mangle]
-pub unsafe extern "C" fn generate_private_key(out_private_key: *mut *mut PrivateKey) -> bool {
-    let mut rng = thread_rng();
-    let key = PrivateKey::generate(&mut rng);
-    *out_private_key = Box::into_raw(Box::new(key));
-
-    true
-}
-
-fn deserialize<T: FromBytes>(in_bytes: *const u8, in_bytes_len: c_int, out: *mut *mut T) -> bool {
-    convert_result_to_bool::<_, std::io::Error, _>(|| {
-        let bytes = unsafe { slice::from_raw_parts(in_bytes, in_bytes_len as usize) };
-        let key = T::read(bytes)?;
-        unsafe {
-            *out = Box::into_raw(Box::new(key));
-        }
-
-        Ok(())
-    })
-}
-
-fn serialize<T: ToBytes>(in_obj: *const T, out_bytes: *mut *mut u8, out_len: *mut c_int) -> bool {
-    convert_result_to_bool::<_, std::io::Error, _>(|| {
-        let obj = unsafe { &*in_obj };
-        let mut obj_bytes = vec![];
-        obj.write(&mut obj_bytes)?;
-        obj_bytes.shrink_to_fit();
-        unsafe {
-            *out_bytes = obj_bytes.as_mut_ptr();
-            *out_len = obj_bytes.len() as c_int;
-        }
-        std::mem::forget(obj_bytes);
-
-        Ok(())
-    })
-}
-
-#[no_mangle]
-pub extern "C" fn deserialize_private_key(
-    in_private_key_bytes: *const u8,
-    in_private_key_bytes_len: c_int,
-    out_private_key: *mut *mut PrivateKey,
-) -> bool {
-    deserialize(
-        in_private_key_bytes,
-        in_private_key_bytes_len,
-        out_private_key,
-    )
-}
-
-#[no_mangle]
-pub extern "C" fn serialize_private_key(
-    in_private_key: *const PrivateKey,
-    out_bytes: *mut *mut u8,
-    out_len: *mut c_int,
-) -> bool {
-    serialize(in_private_key, out_bytes, out_len)
-}
-
-#[no_mangle]
-pub extern "C" fn private_key_to_public_key(
-    in_private_key: *const PrivateKey,
-    out_public_key: *mut *mut PublicKey,
-) -> bool {
-    convert_result_to_bool::<_, std::io::Error, _>(|| {
-        let private_key = unsafe { &*in_private_key };
-        let public_key = private_key.to_public();
-        unsafe {
-            *out_public_key = Box::into_raw(Box::new(public_key));
-        }
-
-        Ok(())
-    })
-}
-
-#[no_mangle]
-pub extern "C" fn sign_message(
-    in_private_key: *const PrivateKey,
-    in_message: *const u8,
-    in_message_len: c_int,
-    in_extra_data: *const u8,
-    in_extra_data_len: c_int,
-    should_use_composite: bool,
-    out_signature: *mut *mut Signature,
-) -> bool {
-    convert_result_to_bool::<_, Box<dyn Error>, _>(|| {
-        let private_key = unsafe { &*in_private_key };
-        let message = unsafe { slice::from_raw_parts(in_message, in_message_len as usize) };
-        let extra_data =
-            unsafe { slice::from_raw_parts(in_extra_data, in_extra_data_len as usize) };
-        let signature = if should_use_composite {
-            private_key.sign(message, extra_data, &*COMPOSITE_HASH_TO_G1)?
-        } else {
-            private_key.sign(message, extra_data, &*DIRECT_HASH_TO_G1)?
-        };
-        unsafe {
-            *out_signature = Box::into_raw(Box::new(signature));
-        }
-
-        Ok(())
-    })
-}
-
-#[no_mangle]
-pub extern "C" fn sign_pop(
-    in_private_key: *const PrivateKey,
-    in_message: *const u8,
-    in_message_len: c_int,
-    out_signature: *mut *mut Signature,
-) -> bool {
-    convert_result_to_bool::<_, Box<dyn Error>, _>(|| {
-        let private_key = unsafe { &*in_private_key };
-        let message = unsafe { slice::from_raw_parts(in_message, in_message_len as usize) };
-        let signature = private_key.sign_pop(message, &*DIRECT_HASH_TO_G1)?;
-        unsafe {
-            *out_signature = Box::into_raw(Box::new(signature));
-        }
-
-        Ok(())
-    })
-}
-
-#[no_mangle]
-pub extern "C" fn hash_direct(
-    in_message: *const u8,
-    in_message_len: c_int,
-    out_hash: *mut *mut u8,
-    out_len: *mut c_int,
-    use_pop: bool,
-) -> bool {
-    convert_result_to_bool::<_, Box<dyn Error>, _>(|| {
-        let message = unsafe { slice::from_raw_parts(in_message, in_message_len as usize) };
-        let domain = if use_pop { POP_DOMAIN } else { SIG_DOMAIN };
-        let hash = DIRECT_HASH_TO_G1.hash(domain, message, &[])?;
-        let mut obj_bytes = vec![];
-        hash.into_affine().write(&mut obj_bytes)?;
-        obj_bytes.shrink_to_fit();
-        unsafe {
-            *out_hash = obj_bytes.as_mut_ptr();
-            *out_len = obj_bytes.len() as c_int;
-        }
-        std::mem::forget(obj_bytes);
-        Ok(())
-    })
-}
-
-#[no_mangle]
-pub extern "C" fn hash_composite(
-    in_message: *const u8,
-    in_message_len: c_int,
-    in_extra_data: *const u8,
-    in_extra_data_len: c_int,
-    out_hash: *mut *mut u8,
-    out_len: *mut c_int,
-) -> bool {
-    convert_result_to_bool::<_, Box<dyn Error>, _>(|| {
-        let message = unsafe { slice::from_raw_parts(in_message, in_message_len as usize) };
-        let extra_data =
-            unsafe { slice::from_raw_parts(in_extra_data, in_extra_data_len as usize) };
-        let hash = COMPOSITE_HASH_TO_G1.hash(SIG_DOMAIN, message, extra_data)?;
-        let mut obj_bytes = vec![];
-        hash.write(&mut obj_bytes)?;
-        obj_bytes.shrink_to_fit();
-        unsafe {
-            *out_hash = obj_bytes.as_mut_ptr();
-            *out_len = obj_bytes.len() as c_int;
-        }
-        std::mem::forget(obj_bytes);
-        Ok(())
-    })
-}
-
-#[no_mangle]
-pub extern "C" fn compress_signature(
-    in_signature: *const u8,
-    in_signature_len: c_int,
-    out_signature: *mut *mut u8,
-    out_len: *mut c_int,
-) -> bool {
-    convert_result_to_bool::<_, Box<dyn Error>, _>(|| {
-        let signature = unsafe { slice::from_raw_parts(in_signature, in_signature_len as usize) };
-        let x = Fq::read(&signature[0..48]).unwrap();
-        let y = Fq::read(&signature[48..96]).unwrap();
-        let affine = G1Affine::new(x, y, false);
-        let sig = Signature::from(affine.into_projective());
-        let mut obj_bytes = vec![];
-        sig.write(&mut obj_bytes)?;
-        obj_bytes.shrink_to_fit();
-        unsafe {
-            *out_signature = obj_bytes.as_mut_ptr();
-            *out_len = obj_bytes.len() as c_int;
-        }
-        std::mem::forget(obj_bytes);
-        Ok(())
-    })
-}
-
-#[no_mangle]
-pub extern "C" fn compress_pubkey(
-    in_pubkey: *const u8,
-    in_pubkey_len: c_int,
-    out_pubkey: *mut *mut u8,
-    out_len: *mut c_int,
-) -> bool {
-    convert_result_to_bool::<_, Box<dyn Error>, _>(|| {
-        let pubkey = unsafe { slice::from_raw_parts(in_pubkey, in_pubkey_len as usize) };
-        let x = Fq2::read(&pubkey[0..96]).unwrap();
-        let y = Fq2::read(&pubkey[96..192]).unwrap();
-        let affine = G2Affine::new(x, y, false);
-        let pk = PublicKey::from(affine.into_projective());
-        let mut obj_bytes = vec![];
-        pk.write(&mut obj_bytes)?;
-        obj_bytes.shrink_to_fit();
-        unsafe {
-            *out_pubkey = obj_bytes.as_mut_ptr();
-            *out_len = obj_bytes.len() as c_int;
-        }
-        std::mem::forget(obj_bytes);
-        Ok(())
-    })
-}
-
-/// # Safety
-///
-/// This function must only be called on a valid PrivateKey instance pointer.
-#[no_mangle]
-pub unsafe extern "C" fn destroy_private_key(private_key: *mut PrivateKey) -> bool {
-    if private_key.is_null() {
-        return false;
-    }
-    Box::from_raw(private_key);
-    true
-}
-
-/// # Safety
-///
-/// This function must only be called on a valid vector pointer.
-#[no_mangle]
-pub unsafe extern "C" fn free_vec(bytes: *mut u8, len: c_int) -> bool {
-    if bytes.is_null() {
-        return false;
-    }
-    Vec::from_raw_parts(bytes, len as usize, len as usize);
-    true
-}
-
-/// # Safety
-///
-/// This function must only be called on a valid PublicKey instance pointer.
-#[no_mangle]
-pub unsafe extern "C" fn destroy_public_key(public_key: *mut PublicKey) -> bool {
-    if public_key.is_null() {
-        return false;
-    }
-    Box::from_raw(public_key);
-    true
-}
-
-/// # Safety
-///
-/// This function must only be called on a valid Signature instance pointer.
-#[no_mangle]
-pub unsafe extern "C" fn destroy_signature(signature: *mut Signature) -> bool {
-    if signature.is_null() {
-        return false;
-    }
-    Box::from_raw(signature);
-    true
-}
-
-#[no_mangle]
-pub extern "C" fn deserialize_public_key(
-    in_public_key_bytes: *const u8,
-    in_public_key_bytes_len: c_int,
-    out_public_key: *mut *mut PublicKey,
-) -> bool {
-    deserialize(in_public_key_bytes, in_public_key_bytes_len, out_public_key)
-}
-
-#[no_mangle]
-pub extern "C" fn serialize_public_key(
-    in_public_key: *const PublicKey,
-    out_bytes: *mut *mut u8,
-    out_len: *mut c_int,
-) -> bool {
-    serialize(in_public_key, out_bytes, out_len)
-}
-
-#[no_mangle]
-pub extern "C" fn deserialize_signature(
-    in_signature_bytes: *const u8,
-    in_signature_bytes_len: c_int,
-    out_signature: *mut *mut Signature,
-) -> bool {
-    deserialize(in_signature_bytes, in_signature_bytes_len, out_signature)
-}
-
-#[no_mangle]
-pub extern "C" fn serialize_signature(
-    in_signature: *const Signature,
-    out_bytes: *mut *mut u8,
-    out_len: *mut c_int,
-) -> bool {
-    serialize(in_signature, out_bytes, out_len)
-}
-
-#[no_mangle]
-pub extern "C" fn verify_signature(
-    in_public_key: *const PublicKey,
-    in_message: *const u8,
-    in_message_len: c_int,
-    in_extra_data: *const u8,
-    in_extra_data_len: c_int,
-    in_signature: *const Signature,
-    should_use_composite: bool,
-    out_verified: *mut bool,
-) -> bool {
-    convert_result_to_bool::<_, std::io::Error, _>(|| {
-        let public_key = unsafe { &*in_public_key };
-        let message = unsafe { slice::from_raw_parts(in_message, in_message_len as usize) };
-        let extra_data =
-            unsafe { slice::from_raw_parts(in_extra_data, in_extra_data_len as usize) };
-        let signature = unsafe { &*in_signature };
-        let verified = if should_use_composite {
-            public_key
-                .verify(message, extra_data, signature, &*COMPOSITE_HASH_TO_G1)
-                .is_ok()
-        } else {
-            public_key
-                .verify(message, extra_data, signature, &*DIRECT_HASH_TO_G1)
-                .is_ok()
-        };
-        unsafe { *out_verified = verified };
-
-        Ok(())
-    })
-}
-
-#[no_mangle]
-/// Receives a list of messages composed of:
-/// 1. the data
-/// 1. the public keys which signed on the data
-/// 1. the signature produced by the public keys
-///
-/// It will create the aggregate signature from all messages and execute batch
-/// verification against each (data, publickey) pair. Internally calls `Signature::batch_verify`
-///
-/// The verification equation can be found in pg.11 from
-/// https://eprint.iacr.org/2018/483.pdf: "Batch verification"
-pub extern "C" fn batch_verify_signature(
-    messages_ptr: *const MessageFFI,
-    messages_len: usize,
-    should_use_composite: bool,
-    verified: *mut bool,
-) -> bool {
-    convert_result_to_bool::<_, BLSError, _>(|| {
-        // Get the pointers slice
-        let messages: &[MessageFFI] = unsafe { slice::from_raw_parts(messages_ptr, messages_len) };
-
-        // Get the data from the underlying pointers in the right format
-        let messages = messages.iter().map(Message::from).collect::<Vec<_>>();
-
-        let asig = Signature::aggregate(messages.iter().map(|m| m.sig));
-
-        let pubkeys = messages.iter().map(|m| m.public_key).collect::<Vec<_>>();
-        let messages = messages
-            .iter()
-            .map(|m| (m.data, m.extra))
-            .collect::<Vec<_>>();
-
-        let is_verified = if should_use_composite {
-            asig.batch_verify(&pubkeys, SIG_DOMAIN, &messages, &*COMPOSITE_HASH_TO_G1)
-                .is_ok()
-        } else {
-            asig.batch_verify(&pubkeys, SIG_DOMAIN, &messages, &*DIRECT_HASH_TO_G1)
-                .is_ok()
-        };
-
-        unsafe { *verified = is_verified };
-        Ok(())
-    })
-}
-
-#[no_mangle]
-pub extern "C" fn verify_pop(
-    in_public_key: *const PublicKey,
-    in_message: *const u8,
-    in_message_len: c_int,
-    in_signature: *const Signature,
-    out_verified: *mut bool,
-) -> bool {
-    convert_result_to_bool::<_, std::io::Error, _>(|| {
-        let public_key = unsafe { &*in_public_key };
-        let message = unsafe { slice::from_raw_parts(in_message, in_message_len as usize) };
-        let signature = unsafe { &*in_signature };
-        let verified = public_key
-            .verify_pop(message, signature, &*DIRECT_HASH_TO_G1)
-            .is_ok();
-        unsafe { *out_verified = verified };
-
-        Ok(())
-    })
-}
-
-#[no_mangle]
-pub extern "C" fn aggregate_public_keys(
-    in_public_keys: *const *const PublicKey,
-    in_public_keys_len: c_int,
-    out_public_key: *mut *mut PublicKey,
-) -> bool {
-    convert_result_to_bool::<_, std::io::Error, _>(|| {
-        let public_keys_ptrs =
-            unsafe { slice::from_raw_parts(in_public_keys, in_public_keys_len as usize) };
-        let public_keys = public_keys_ptrs
-            .to_vec()
-            .into_iter()
-            .map(|pk| unsafe { &*pk }.clone())
-            .collect::<Vec<PublicKey>>();
-        let aggregated_public_key = PublicKeyCache::aggregate(&public_keys[..]);
-        unsafe {
-            *out_public_key = Box::into_raw(Box::new(aggregated_public_key));
-        }
-
-        Ok(())
-    })
-}
-
-#[no_mangle]
-pub extern "C" fn aggregate_public_keys_subtract(
-    in_aggregated_public_key: *const PublicKey,
-    in_public_keys: *const *const PublicKey,
-    in_public_keys_len: c_int,
-    out_public_key: *mut *mut PublicKey,
-) -> bool {
-    convert_result_to_bool::<_, std::io::Error, _>(|| {
-        let aggregated_public_key = unsafe { &*in_aggregated_public_key };
-        let public_keys_ptrs =
-            unsafe { slice::from_raw_parts(in_public_keys, in_public_keys_len as usize) };
-        let public_keys = public_keys_ptrs
-            .to_vec()
-            .into_iter()
-            .map(|pk| unsafe { &*pk }.clone())
-            .collect::<Vec<PublicKey>>();
-        let aggregated_public_key_to_subtract = PublicKeyCache::aggregate(&public_keys[..]);
-        let prepared_aggregated_public_key = PublicKey::from(
-            *aggregated_public_key.as_ref() - *aggregated_public_key_to_subtract.as_ref(),
-        );
-        unsafe {
-            *out_public_key = Box::into_raw(Box::new(prepared_aggregated_public_key));
-        }
-
-        Ok(())
-    })
-}
-
-#[no_mangle]
-pub extern "C" fn aggregate_signatures(
-    in_signatures: *const *const Signature,
-    in_signatures_len: c_int,
-    out_signature: *mut *mut Signature,
-) -> bool {
-    convert_result_to_bool::<_, std::io::Error, _>(|| {
-        let signatures_ptrs =
-            unsafe { slice::from_raw_parts(in_signatures, in_signatures_len as usize) };
-        let signatures = signatures_ptrs
-            .to_vec()
-            .into_iter()
-            .map(|sig| unsafe { &*sig }.clone())
-            .collect::<Vec<Signature>>();
-        let aggregated_signature = Signature::aggregate(&signatures[..]);
-        unsafe {
-            *out_signature = Box::into_raw(Box::new(aggregated_signature));
-        }
-
-        Ok(())
-    })
-}
-
-fn convert_result_to_bool<T, E: Display, F: Fn() -> Result<T, E>>(f: F) -> bool {
-    if let Err(e) = f() {
-        error!("BLS library error: {}", e.to_string());
-        return false;
-    }
-    true
 }
